+++ src/main/kotlin/rat/poison/scripts/aim/FlatAim.kt
@@ -5,5 +5,5 @@ import rat.poison.utils.generalUtil.strToBool
 import rat.poison.utils.writeAim
 
 fun flatAim() = aimScript(curSettings["AIM_DURATION"].toInt(), { curSettings["ENABLE_FLAT_AIM"].strToBool() }) { dest, current, aimSpeed, _ ->
-	writeAim(current, dest, aimSpeed.toDouble())
+	writeAim(current, dest, aimSpeed.toFloat())
 }
